Reduced memory footage even more
Though, it doesnt look really impressive in terms of reducing.

+++ src/me/oak/imgcolor/Bag.java
@@ -9,5 +9,5 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor public class Bag<T> {
 
     public final T value;
-    public int amount;
+    public short amount;
 }
